todo with firebase

+++ Firebase todo app/firebase.js
@@ -1,6 +1,7 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "https://www.gstatic.com/firebasejs/11.9.1/firebase-app.js";
-import { getFirestore, collection, addDoc, getDocs,doc, deleteDoc, updateDoc } from "https://www.gstatic.com/firebasejs/11.9.1/firebase-firestore.js";
+import { getFirestore, collection, addDoc, getDocs,doc,setDoc, deleteDoc, updateDoc } from "https://www.gstatic.com/firebasejs/11.9.1/firebase-firestore.js";
+import { getAuth ,signInWithEmailAndPassword, createUserWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/11.9.1/firebase-auth.js";
 
   // TODO: Add SDKs for Firebase products that you want to use
   // https://firebase.google.com/docs/web/setup#available-libraries
@@ -17,6 +18,7 @@ import { getFirestore, collection, addDoc, getDocs,doc, deleteDoc, updateDoc } f
 
   // Initialize Firebase
 const app = initializeApp(firebaseConfig);
-const db = getFirestore(app)
+const db = getFirestore(app);
+const auth = getAuth();
 
-export{ app, db, collection, addDoc, getDocs,doc, deleteDoc, updateDoc }
+export{ app, db, auth,signInWithEmailAndPassword, createUserWithEmailAndPassword, collection, addDoc, getDocs,doc,setDoc, deleteDoc, updateDoc }
